[wip] change fdp-dynamic-page projection strategy
