+++ apps/docs/src/app/platform/component-docs/platform-dynamic-page/platform-dynamic-page-examples/platform-dynamic-page-snap-scroll-example.component.scss
@@ -8,7 +8,3 @@
     background-color: rgb(255, 255, 255);
     overflow-x: hidden;
 }
-
-.overlay-content {
-    width: 100%;
-}
