--- libs/platform/src/lib/components/dynamic-page/dynamic-page-content/dynamic-page-content.component.html
@@ -1,8 +1,0 @@
-<!-- if not tab content, just display normally -->
-<ng-container *ngIf="!tabLabel">
-    <ng-container *ngTemplateOutlet="contentTemplate"></ng-container>
-</ng-container>
-
-<!-- if tab content, get the contentTemplate for passing it on to each individual tab -->
-<ng-template #contentTemplate><ng-content></ng-content></ng-template>
-<div class="footer-spacer"></div>
