+++ libs/platform/src/lib/components/dynamic-page/dynamic-page-header/header/dynamic-page-header-internal.component.html
@@ -9,7 +9,7 @@
     >
         <ng-content></ng-content>
     </div>
-    <div class="fd-dynamic-page__collapsible-header-visibility-container" #pincollapseContainer>
+    <div class="fd-dynamic-page__collapsible-header-visibility-container">
         <ng-container *ngIf="collapsible">
             <div class="fd-dynamic-page__collapsible-header-visibility-container--left-gradient"></div>
             <div
