+++ libs/platform/src/lib/components/dynamic-page/public_api.ts
@@ -6,7 +6,6 @@ export * from './dynamic-page-header/header/dynamic-page-header.component';
 export * from './dynamic-page-header/key-info/dynamic-page-key-info.component';
 export * from './dynamic-page-header/title/dynamic-page-title.component';
 export * from './dynamic-page-content/dynamic-page-content.component';
-export * from './dynamic-page-content/dynamic-page-tabbed-content.component';
 export * from './dynamic-page-footer/dynamic-page-footer.component';
 export * from './dynamic-page.config';
 export * from './dynamic-page.module';
